Cloud Run (#71)
* Migrate app check function to cloud run server

* Update open api documentation

* IEize

* Include test dependency for content negotiation

* Modify open api configuration

* Resolve mismatched ktor client/server dependencies after removing shadowJar occlusion

+++ app-check/app-check-sdk/src/main/kotlin/io/ashdavies/check/AppCheckGenerator.kt
@@ -45,7 +45,7 @@ internal fun AppCheckGenerator(
 
     val ttlMillis = result.ttl
         .substring(0, result.ttl.length - 1)
-        .toInt() * 1000
+        .toLong() * 1000
 
     Response.Normalised(
         ttlMillis = ttlMillis,
